requirements.txt
fastapi
uvicorn
pandas
langchain
langchain-community
langchain-core
langchainhub
langchain-google-genai
python-dotenv
dnspython
langgraph
python-multipart
certifi
supabase
google-generativeai
pandasql
openpyxl
duckdb
matplotlib
pyjwt
bcrypt
python-jose[cryptography]
passlib[bcrypt]
python-json-logger
requests